Show first map

+++ src/board.h
@@ -11,18 +11,18 @@
 class Board : public QObject
 {
     Q_OBJECT
-    Q_PROPERTY(int width READ width WRITE setWidth NOTIFY widthChanged)
-    Q_PROPERTY(int height READ height WRITE setHeight NOTIFY heightChanged)
+    Q_PROPERTY(qreal width READ width WRITE setWidth NOTIFY widthChanged)
+    Q_PROPERTY(qreal height READ height WRITE setHeight NOTIFY heightChanged)
     Q_PROPERTY(QQmlListProperty<Tile> tiles READ tiles CONSTANT)
 public:
     explicit Board(QObject *parent = nullptr);
     ~Board();
 
-    int width() const { return width_; }
-    void setWidth(int width);
+    qreal width() const { return width_; }
+    void setWidth(qreal width);
 
-    int height() const { return height_; }
-    void setHeight(int height);
+    qreal height() const { return height_; }
+    void setHeight(qreal height);
 
     Tile *tile(int row, int column) const;
     QQmlListProperty<Tile> tiles();
@@ -39,8 +39,8 @@ private:
     void initialize();
     void fillTiles();
 
-    int width_ = 0;
-    int height_ = 0;
+    qreal width_ = 0;
+    qreal height_ = 0;
     QList<Tile *> tiles_;
 
     static const int SIZE;
